Managed to hit the 50% coverage
- Would like to test controllers too, those will cover a lot as well.
- Security, config and some other auth and static classes are not well tested, i hope to get around to those later

+++ src/main/java/com/willpowered/eindprojectwpsbe/Project/ProjectService.java
@@ -200,7 +200,11 @@ public class ProjectService {
     //// Delete
 
     public void deleteProject(Long id) {
-        projectRepository.deleteById(id);
+        var optionalProject = projectRepository.findById(id);
+        if (optionalProject.isPresent()) {
+            projectRepository.deleteById(id);
+        } else {
+            throw new RecordNotFoundException("Project does not exist");
+        }
     }
-
 }
